banco_de_dados_p2-main/requirements.txt
fastapi
uvicorn[standard]
motor
redis
aio-pika
pydantic
python-dotenv
